EX-2465 Catch errors thrown by operators
In formulas, infix operators are handled by the formula-parser package,
whereas functions are handled by formula.js. When an `Error` is passed
into an infix operation, it was not being handled.

Errors passed into infix operators (`+`, `-`, `&`, etc.) were getting
incorporated into the result (so `&`, which concatenates, would give
results like 'abc{ Error: ... }'). This commit changes the calling code
to first check the parameters for an `Error` instance and, if found,
throw that error. It is then handled the same as any error from a
function imported from formula.js.

Removes tests for bessel functions (which we don't support and have been
removed from our fork of formua.js).

Updates a few tests that were failing on near-matches to use the
"close-enough" check `toBeMatchCloseTo`.

Adds explicit tests for the event emitter used to handle function calls
and function errors when formulas are being calculated.